Changed references
Previous references included rfc for RADIUS and docs for TACACS+, both of which are network authentication/management protocols and have nothing to do with command line monitoring.  I've included references to Sysmon for windows, auditd command-line push to rsyslog for linux, and and instructional blog for auditing with OpenBSM, which is included natively in MacOS since Snow Leopard.